Improve layout on mobile devices.
Chrome for mobile in particular enlarges some of the text in an
attempt to make the site more readable, but this generally gets
the wrong part of the page (e.g. making footnotes way bigger than
the main text). Disable this "font boosting" by setting max-height
on the parent element.