Mark certain pages as 'responsive'.
Pages so marked will get an HTML viewport meta tag set to tell
mobile devices to fit them within the device width, rather than
trying to pretend they're normal web pages (and forcing the user
to read tiny text or to scroll the screen a lot). This only works
right now for the help pages, which generally can be resized to
fit in a mobile device's screen without looking terrible.