Add a new FINALIZING state.
This new state comes between POSTPROCESSING and COMPLETED,
and can be used for any special processing needed at the
last step of a multi-run job. For example, AllosMod-FoXS
runs FoXS after the last step, and it currently does this
in the complete() method. This isn't a good place for it,
since end users see that their job is reporting that it's
finished and try to download results before the FoXS run
is done. postprocess() isn't a good place either, because
job outputs haven't been moved from network to local
storage at this point. This FoXS run can now go in the
FINALIZING state.