Test Perl code in its original location.
Previously since the web service Perl code was
not valid Perl (@CONFIG@ had to be substituted at
build time) in order to test it we had to install it
in a temporary location. This confuses Perl coverage
tools, since they use the wrong paths (or can't find
the files at all). Replace @CONFIG@ with
"##CONFIG##" which is at least valid syntax so we can
test in the original location.