Report an error for too-long form fields
Notify the user if any form field exceeds its
maximum length. This shouldn't happen in regular
use since the fields have a suitable max length set
in the input HTML forms, but a manually-constructed
POST could do it. Report a (terse) error rather than
failing with an internal server error (which also
bothers the admin with an email).